Implement event sourcing model rather than instance variables
I decided to do this because I'm still not sure if the update sidebar
method (in particular) should actually be in the Creature class. Now
though, I can quickly move it out and not have to change any of the
calling code.

I've no idea if that's a good reason.